requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.8.0
django-jazzmin==3.0.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kavenegar==1.1.2
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
